Refactored the treasury pallet
