+++ runtime/src/configs/mod.rs
@@ -26,3 +26,6 @@ pub use grandpa::*;
 
 pub mod system;
 pub use system::*;
+
+pub mod treasury;
+pub use treasury::*;
